Implement a new plugin system using multiple-inheritance
First, we split all the interesting functions out of Yap into a new
class, YapCore.  Yap retains only the "main" method.  Yap is a metaclass
with dynamically-determined inheritance.  The yap_metaclass function
loads all plugins and creates the list of parents of Yap.  In this way,
commands (and even utility functions) can be overriden in a sane way
using standard python OO.